backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "pillow>=12.0.0",
    "python-decouple>=3.8",
]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]
